Create Public-Key-Encryption
An Assignment I did for Computer Security class during my Undergrad at Georgia Southern 
University. It has an interactive mode as well as a command prompt mode 
that allows you to encrypt a file and choose the key length for the encryption. It then 
generates the Public and Private keys as well as the encrypted text file. 